# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/subscriptions/status": {
      "get": {
        "operationId": "SubscriptionsController_status",
        "summary": "Get a user's current subscription plan",
        "description": "Defaults to the free plan if the user has never subscribed.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Any non-owner role may manage its own subscription.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "example": 4,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionDto"
                }
              }
            }
          }
        },
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/api/subscriptions/upgrade": {
      "post": {
        "operationId": "SubscriptionsController_upgrade",
        "summary": "Upgrade (or change) a subscription plan",
        "description": "Free → Plus → Ultra Pro. Logs a simulated payment for paid plans.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Any non-owner role may manage its own subscription.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradeSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The new active subscription record"
          }
        },
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/api/subscriptions/cancel": {
      "post": {
        "operationId": "SubscriptionsController_cancel",
        "summary": "Cancel a subscription",
        "description": "Reverts the user to the free plan.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Any non-owner role may manage its own subscription.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Subscription for user 4 reverted to free plan"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/api/organisers/apply": {
      "post": {
        "operationId": "OrganisersController_apply",
        "summary": "Apply to become a Verified Organizer",
        "description": "Creates a pending application. Does not itself grant the organizer role — see PATCH /:id.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. apply/profile/analytics/subscription-upgrade: user | organizer | admin. PATCH :id (verify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyOrganiserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "organisers"
        ]
      }
    },
    "/api/organisers/{id}": {
      "patch": {
        "operationId": "OrganisersController_setStatus",
        "summary": "Verify or reject an organiser application",
        "description": "Admin-only. On verification, also flips the applicant's account role to organizer.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. apply/profile/analytics/subscription-upgrade: user | organizer | admin. PATCH :id (verify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganiserStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "organisers"
        ]
      }
    },
    "/api/organisers/profile": {
      "get": {
        "operationId": "OrganisersController_profile",
        "summary": "Get a user's organiser application/plan status",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. apply/profile/analytics/subscription-upgrade: user | organizer | admin. PATCH :id (verify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "example": 6,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Application status, plan, and plan limits"
          }
        },
        "tags": [
          "organisers"
        ]
      }
    },
    "/api/organisers/analytics": {
      "get": {
        "operationId": "OrganisersController_analytics",
        "summary": "Get an organiser's tournament analytics",
        "description": "Aggregates their own events/registrations — same scoping as the Organizer Dashboard.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. apply/profile/analytics/subscription-upgrade: user | organizer | admin. PATCH :id (verify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "example": 6,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "organisers"
        ]
      }
    },
    "/api/organisers/subscription/upgrade": {
      "post": {
        "operationId": "OrganisersController_upgradePlan",
        "summary": "Upgrade the organiser subscription tier",
        "description": "Free Organizer → Premium Organizer. Logs a simulated payment.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. apply/profile/analytics/subscription-upgrade: user | organizer | admin. PATCH :id (verify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradeOrganiserPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "organisers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gameunity Backend API",
    "description": "In-memory NestJS backend for frontend integration and academic evaluation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Dev Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "RBAC role header: admin | community_manager | moderator | user"
      }
    },
    "schemas": {
...
      "SubscriptionDto": {
        "type": "object",
        "properties": {
          "plan": {
            "type": "string",
            "example": "plus",
            "enum": [
              "free",
              "plus",
              "ultra_pro"
            ]
          },
          "status": {
            "type": "string",
            "example": "active",
            "enum": [
              "active",
              "cancelled"
            ]
          },
          "startedAt": {
            "type": "string",
            "example": "2026-05-01T09:00:00.000Z"
          }
        },
        "required": [
          "plan",
          "status"
        ]
      },
      "UpgradeSubscriptionDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 4
          },
          "plan": {
            "type": "string",
            "example": "plus",
            "enum": [
              "free",
              "plus",
              "ultra_pro"
            ]
          }
        },
        "required": [
          "userId",
          "plan"
        ]
      },
      "CancelSubscriptionDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 4
          }
        },
        "required": [
          "userId"
        ]
      },
      "ApplyOrganiserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 4
          },
          "experienceNote": {
            "type": "string",
            "example": "Ran 3 campus BGMI tournaments in the last year."
          }
        },
        "required": [
          "userId"
        ]
      },
      "UpdateOrganiserStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "verified",
            "enum": [
              "verified",
              "rejected"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "UpgradeOrganiserPlanDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 6
          },
          "plan": {
            "type": "string",
            "example": "premium",
            "enum": [
              "free",
              "premium"
            ]
          }
        },
        "required": [
          "userId",
          "plan"
        ]
      },
...
    }
  }
}